move JsonSchemaValidator.java into commons
