+++ dataline-analytics/src/main/java/io/dataline/analytics/TrackingClientSingleton.java
@@ -29,7 +29,7 @@ import io.dataline.config.Configs;
 import io.dataline.config.StandardWorkspace;
 import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigRepository;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import io.dataline.config.persistence.PersistenceConstants;
 import java.io.IOException;
 import java.util.function.Supplier;
