+++ dataline-commons/src/main/java/io/dataline/commons/json/JsonValidationException.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package io.dataline.config.persistence;
+package io.dataline.commons.json;
 
 public class JsonValidationException extends Exception {
