+++ dataline-config/persistence/build.gradle
@@ -1,9 +1,5 @@
 dependencies {
     implementation group: 'commons-io', name: 'commons-io', version: '2.7'
 
-    implementation group: 'com.networknt', name: 'json-schema-validator', version: '1.0.42'
-    // needed so that we can follow $ref when parsing json. jackson does not support this natively.
-    implementation 'me.andrz.jackson:jackson-json-reference-core:0.3.2'
-
     implementation project(':dataline-config:models')
 }
