+++ dataline-config/persistence/src/main/java/io/dataline/config/persistence/ConfigPersistence.java
@@ -24,6 +24,7 @@
 
 package io.dataline.config.persistence;
 
+import io.dataline.commons.json.JsonValidationException;
 import io.dataline.config.ConfigSchema;
 import java.io.IOException;
 import java.util.List;
