+++ dataline-scheduler/src/main/java/io/dataline/scheduler/job_factory/DefaultSyncJobFactory.java
@@ -27,7 +27,7 @@ package io.dataline.scheduler.job_factory;
 import io.dataline.config.StandardSync;
 import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigRepository;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import io.dataline.scheduler.persistence.SchedulerPersistence;
 import java.io.IOException;
 import java.util.UUID;
