+++ dataline-server/src/main/java/io/dataline/server/handlers/ConnectionsHandler.java
@@ -41,7 +41,7 @@ import io.dataline.config.StandardSync;
 import io.dataline.config.StandardSyncSchedule;
 import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigRepository;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import io.dataline.server.converters.SchemaConverter;
 import java.io.IOException;
 import java.util.Collections;
