+++ dataline-server/src/main/java/io/dataline/server/handlers/DestinationImplementationsHandler.java
@@ -37,7 +37,7 @@ import io.dataline.config.DestinationConnectionSpecification;
 import io.dataline.config.StandardDestination;
 import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigRepository;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import io.dataline.server.validation.IntegrationSchemaValidation;
 import java.io.IOException;
 import java.util.List;
