+++ dataline-server/src/main/java/io/dataline/server/handlers/DestinationSpecificationsHandler.java
@@ -29,7 +29,7 @@ import io.dataline.api.model.DestinationSpecificationRead;
 import io.dataline.config.DestinationConnectionSpecification;
 import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigRepository;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import io.dataline.server.errors.KnownException;
 import java.io.IOException;
 import org.eclipse.jetty.http.HttpStatus;
