+++ dataline-server/src/main/java/io/dataline/server/handlers/SchedulerHandler.java
@@ -42,7 +42,7 @@ import io.dataline.config.StandardDiscoverSchemaOutput;
 import io.dataline.config.StandardSync;
 import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigRepository;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import io.dataline.scheduler.Job;
 import io.dataline.scheduler.JobStatus;
 import io.dataline.scheduler.persistence.SchedulerPersistence;
