+++ dataline-server/src/main/java/io/dataline/server/handlers/SourceSpecificationsHandler.java
@@ -29,7 +29,7 @@ import io.dataline.api.model.SourceSpecificationRead;
 import io.dataline.config.SourceConnectionSpecification;
 import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigRepository;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import io.dataline.server.errors.KnownException;
 import java.io.IOException;
