+++ dataline-server/src/main/java/io/dataline/server/handlers/SourcesHandler.java
@@ -30,7 +30,7 @@ import io.dataline.api.model.SourceReadList;
 import io.dataline.config.StandardSource;
 import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigRepository;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import java.io.IOException;
 import java.util.List;
 import java.util.stream.Collectors;
