+++ dataline-server/src/main/java/io/dataline/server/handlers/WebBackendConnectionsHandler.java
@@ -37,7 +37,7 @@ import io.dataline.api.model.WbConnectionReadList;
 import io.dataline.api.model.WorkspaceIdRequestBody;
 import io.dataline.commons.enums.Enums;
 import io.dataline.config.persistence.ConfigNotFoundException;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import java.io.IOException;
 import java.util.List;
