+++ dataline-server/src/main/java/io/dataline/server/handlers/WorkspacesHandler.java
@@ -32,7 +32,7 @@ import io.dataline.api.model.WorkspaceUpdate;
 import io.dataline.config.StandardWorkspace;
 import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigRepository;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import io.dataline.config.persistence.PersistenceConstants;
 import java.io.IOException;
 import java.util.UUID;
