+++ dataline-server/src/test/java/io/dataline/server/handlers/DestinationsHandlerTest.java
@@ -35,7 +35,7 @@ import io.dataline.api.model.DestinationReadList;
 import io.dataline.config.StandardDestination;
 import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigRepository;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import java.io.IOException;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
