+++ dataline-server/src/test/java/io/dataline/server/handlers/SourceImplementationsHandlerTest.java
@@ -49,7 +49,7 @@ import io.dataline.config.StandardSource;
 import io.dataline.config.StandardSync;
 import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigRepository;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import io.dataline.server.helpers.ConnectionHelpers;
 import io.dataline.server.helpers.SourceHelpers;
 import io.dataline.server.helpers.SourceImplementationHelpers;
