+++ dataline-server/src/test/java/io/dataline/server/handlers/SourcesHandlerTest.java
@@ -35,7 +35,7 @@ import io.dataline.api.model.SourceReadList;
 import io.dataline.config.StandardSource;
 import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigRepository;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import java.io.IOException;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
