+++ dataline-server/src/test/java/io/dataline/server/handlers/WorkspacesHandlerTest.java
@@ -36,7 +36,7 @@ import io.dataline.api.model.WorkspaceUpdate;
 import io.dataline.config.StandardWorkspace;
 import io.dataline.config.persistence.ConfigNotFoundException;
 import io.dataline.config.persistence.ConfigRepository;
-import io.dataline.config.persistence.JsonValidationException;
+import io.dataline.commons.json.JsonValidationException;
 import io.dataline.config.persistence.PersistenceConstants;
 import java.io.IOException;
 import java.util.UUID;
